Switch labels and field values on change of source or target language
This adds UI functionality to react to the user changing either the
source or target language widgets.

* When the source is changed, the labels of the translate form change
  to the desired language.
* When the target is changed, the translate form's field values
  change to the desired language.

A TranslateController test is also added so that this patch doesn't
fail because of decreased code coverage. This test doesn't actually
test the functionality introduced by this patch because this is all
front-end behaviour (and we don't have Selenium tests or similar).

The built-in Symfony translation package is also disabled, in order
for the HTTP-based controller test to pass.

Bug: https://phabricator.wikimedia.org/T207709