Switch to PNG rendering for image display
This patch changes the API to return PNG images instead of SVG,
so that we can specify the language that is used for the rendering.
To do this, we pass the name of the cached SVG file to a call to
the external rsvg-convert command, which returns the PNG to stdout.
This is served in the usual way.

This change also breaks the API URLs by adding in a 'lang' URL
parameter. This is required because we may not be rendering to the
tool user's current interface language.

When a POST request is given with new translations, the resulting
changed SVG file has to be written to the filesystem so that
rsvg-convert can access it. This is done by creating a new unique
temp file in the same images' cache directory, so that the files
thus created can be cleaned up by the existing clearing
mechanism.

The Docker image we're using doesn't (until T151656 is resolved)
have rsvg-convert installed, so this patch also adds this to the
dockerstart script.

Bug: https://phabricator.wikimedia.org/T211637