Image preview (on load), documentation, and better tests.
This fixes the URL-generation for the preview image, handing it
off to the existing system that fires as the user changes the
form (it replicates what would happen e.g. if a user swiched
languages and then modified a form field). This means we don't
have to generate a URL, which makes things a bit simpler.

Some documentation and tests are also extended, to hopefully
make things a bit clearer -- although there's lots more to be
done in that area.

Bug: https://phabricator.wikimedia.org/T207203