Use separate OOUI stylesheets instead of minified one
This removes oojs-ui-wikimediaui.min.css and replaces it with
the core, widgets, and windows stylesheets because these are
smaller and include (almost) all the bits we need. This is to
stop Webpack failing on the larger file, and to just make things
more efficient.

The 'almost' is because indicators are handled separately, and
so this patch also adds a manual inclusion of the 'down'
indicator in our app.less file.

An out-of-date comment is also removed.

Bug: T217393