Use different no-translations message icon
Swich to a grey Font Awesome question-mark icon for the message
that's shown when an SVG has nothing to translate.

The icon's SVG file is modified to be the correct colour.

A Twig file is used instead of an OOUI object because we no
longer need to use OOUI for this message and it's easier to
load image files from Twig.

Bug: T210218