[EXPERIMENTAL] Add RTL support with CSSJanus webpack plugin
This is experimental, and is attempting to provide RTL support through
the CSSJanus plugin.

There are several issues to be fixed, and potentially they may involve
redoing some of the entrypoints. While RTL support seems to be working
with this out of the box, there might be better ways to do this by
making the external libraries (OOUI especially) their own entrypoint
since they have their own RTL version already-made in their repos,
and then attaching those files independently to the base.html.twig.

This is experimental, to see if it's working properly and covers
the necessary cases, and can serve as a testbed for seeing whether
it is worth expanding to our other tools.

If/when we can examine this and determine it's worth using, we
can remove the experimental tag and merge.

Reference: https://phabricator.wikimedia.org/T208371