Use bundle's login redirection instead of custom one
This switches to using the ToolforgeBundle's redirection feature
instead of a custom session-based one. This requires a new OAuth
consumer with "Allow consumer to specify a callback in requests"
enabled so we can submit the redirect URL.

All PHP dependencies are also updated (apart from OOUI, which
might require front-end changes as well).

https://phabricator.wikimedia.org/T223769

Bug: T223769