Create (and catch) exceptions for all existing error states
Instead of just returning false in SvgFile::makeTranslationReady()
(the return value of which was not being used anyway), make each
of the error states throw an exception. The TranslateController
then catches these, and displays a useful message to the user,
hopefully making it easier for them to fix the SVG.

This also removes the NestedTspanException, because it's not anything
special compared to the other, and I didn't want to go and make
a dozen new exception classes (although in some ways that might've
been better).

There are more tests that could be added for all this, but for now
I think this is sufficient (this change isn't actually doing anything
with the error states, it's just surfacing them better; if there are
incorrect states then those can be followed up separately).

Bug: T271000