Fail earlier for multiple-text-same-lang structure
It's not permitted to have multiple `<text>` elements with the same language
code in a `<switch>` element. This wasn't being checked in
`SvgFile::makeTranslationRead()`, but only later in
`SvgFile::switchToTranslationSet()`, and so it wasn't being reported to the
user when they opened a file for translation, but only later when they tried to
load a translated version (at which point it failed without telling them why).

This change moves the multiple-language check to the earlier method, and fixes
up some broken parts of 91d30977e1180e983b3720c3aa486f8d88dae29d such as the
error message name and parameters, as well as i18n messages and documentation.

This doesn't remove the later check, even though it's now redundant. That can
be done in a follow-up.

Bug: T319154