Target PHP 8.2 and above, and upgrade PHPUnit to version 9
Update CI, Composer, and PHPUnit configuration.

Remove the Maker bundle, as it's not really needed now we're in
maintenanace phase and it means we don't have to upgrade a bunch
of other things to Symfony 6 or 7 right now. We can add it back
in later after the rest of the upgrade if it's still wanted.

Bug: T357844